subha: changed data type of some of the configuration variables to unsigned int. These were wrongly treated as plain int.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/buildQ@3910 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 584ae267659248264c30d6d54561b0fcde8adb4a

+++ basecode/main.cpp
@@ -132,10 +132,10 @@ void checkChildren( Id parent, const string& info )
 
 Id init( int argc, char** argv, bool& doUnitTests, bool& doRegressionTests )
 {
-	int numCores = getNumCores();
-	int numThreads = numCores;
-	int numNodes = 1;
-	int myNode = 0;
+	unsigned int numCores = getNumCores();
+	unsigned int numThreads = numCores;
+	unsigned int numNodes = 1;
+	unsigned int myNode = 0;
 	bool isSingleThreaded = 0;
 	bool isInfinite = 0;
 	int opt;
@@ -166,11 +166,11 @@ Id init( int argc, char** argv, bool& doUnitTests, bool& doRegressionTests )
 				isInfinite = 1;
 				break;
 			case 'n': // Multiple nodes
-				numNodes = atoi( optarg );
+			  numNodes = (unsigned int)atoi( optarg );
 				break;
 			case 't': // Number of process threads to use. Default is
 				// obtained from hardware NumCores.
-				numThreads = atoi( optarg );
+			  numThreads = (unsigned int)atoi( optarg );
 				break;
 			case 'b': // Benchmark: handle later.
 				break;
